refactor(wrapper): Move offering endpoint to use-case api
