+++ extensions/wrapper/wrapper/src/main/java/de/sovity/edc/ext/wrapper/api/usecase/UseCaseResource.java
@@ -14,25 +14,62 @@
 
 package de.sovity.edc.ext.wrapper.api.usecase;
 
+import java.util.List;
+
+import de.sovity.edc.ext.wrapper.api.usecase.model.CreateOfferingDto;
+import de.sovity.edc.ext.wrapper.api.usecase.services.OfferingService;
 import de.sovity.edc.ext.wrapper.api.usecase.model.KpiResult;
 import de.sovity.edc.ext.wrapper.api.usecase.services.KpiApiService;
 import de.sovity.edc.ext.wrapper.api.usecase.services.SupportedPolicyApiService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.GET;
+import jakarta.ws.rs.POST;
 import jakarta.ws.rs.Path;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.core.Response.Status;
 import lombok.RequiredArgsConstructor;
 
-import java.util.List;
 
+/**
+ * Provides the endpoints for use-case specific requests.
+ *
+ * @author Ronja Quensel (ronja.quensel@isst.fraunhofer.de)
+ * @author Richard Treier
+ * @author Tim Dahlmanns
+ */
 @Path("wrapper/use-case-api")
 @Tag(name = "Use Case", description = "Generic Use Case Application API Endpoints.")
 @RequiredArgsConstructor
+@Produces(MediaType.APPLICATION_JSON)
 public class UseCaseResource {
     private final KpiApiService kpiApiService;
     private final SupportedPolicyApiService supportedPolicyApiService;
+    /** Service for managing offerings. */
+    private final OfferingService offeringService;
+
+    /**
+     * Creates a new usecase consisting of asset, policy definition and contract definition.
+     *
+     * @param dto contains all required information for the usecase.
+     * @return a 204 response, if creating the usecase was successful.
+     */
+    @POST
+    @Path("contract-offer")
+    @Consumes(MediaType.APPLICATION_JSON)
+    @Operation(description = "Creates an usecase")
+    public Response createOfferingEndpoint(CreateOfferingDto dto) {
+        if (dto == null) {
+            String error = "No CreateOfferingDto provided";
+            return Response.status(Status.BAD_REQUEST).entity(error).build();
+        }
+
+        offeringService.create(dto);
+        return Response.status(Response.Status.NO_CONTENT).build();
+    }
 
     @GET
     @Path("kpis")
