+++ extensions/wrapper/wrapper/src/test/resources/usecase/contract-offer-valid.json
